// backend/src/utils/enums/auth.enum.ts
export enum AUTH {
  CREATE_USER = "/register-user",
  CREATE_CHEF = "/register-chef",
  ADMIN_SIGNUP = "/admin-signup",
  VERIFY_EMAIL = "/verify-email",
  RESEND_VERIFICATION = "/resend-verification",
  VERIFY_TOKEN = "/verify-token",
  VERIFY_ACCOUNT_TOKEN = "/verify-account-token",
  LOGIN_USER = "/login-user",
  LOGIN_ADMIN = "/login-admin",
  GOOGLE_LOGIN = "/google-login",
  RESEND_TOKEN = "/resend-token",
  RESET_PASSWORD = "/reset-password",
  VERIFY_RESET_TOKEN = "/verify-reset-token",
  FORGOT_PASSWORD = "/forget-password",
  CHANGE_PASSWORD = "/change-password",
  PROFILE = "/profile",
  DELETE_ACCOUNT = "/delete-account",
  LOGOUT = "/logout",
}